Move code dealing with cobra.Command objects into cmd/interal/cli/
