+++ cmd/internal/cli/capability_add_linux.go
