+++ cmd/internal/cli/capability_list_linux.go
