+++ cmd/internal/cli/inspect_linux.go
