+++ cmd/internal/cli/instance_stop_linux.go
