+++ cmd/internal/cli/keys_newpair.go
