+++ cmd/internal/cli/keys_search.go
