+++ cmd/internal/cli/search.go
