+++ cmd/internal/cli/singularity.go
