+++ cmd/internal/cli/verify.go
