+++ cmd/man_pages/man_pages.go
@@ -9,7 +9,7 @@ import (
 	"os"
 
 	"github.com/spf13/cobra/doc"
-	cli "github.com/sylabs/singularity/internal/app/singularity"
+	"github.com/sylabs/singularity/cmd/internal/cli"
 	"github.com/sylabs/singularity/internal/pkg/sylog"
 	"golang.org/x/sys/unix"
 )
